requirements.txt
backoff==1.3.2
boto==2.48.0
cffi==1.8.3
cryptography==1.5.2
Django==1.9.8
django-filter==0.14.0
django-fixtureless==1.4.3.3
django-genes==0.3
django-haystack==2.5.0
django-organisms==0.3
django-ses==0.8.2
django-storages==1.6.3
djangorestframework==3.4.0
djangorestframework-expander==0.2.1
drf-dynamic-fields==0.1.1
enum-compat==0.0.2
eventlet==0.20.1
gevent==1.2.1
greenlet==0.4.11
gunicorn==19.6.0
idna==2.1
psycopg2==2.7.1
pyasn1==0.1.9
pycparser==2.16
pycrypto==2.6.1
PyJWT==1.4.2
requests==2.12.3
six==1.10.0